Added authorisation for delete and edit

+++ backend/routes/posts.js
@@ -7,12 +7,20 @@ const checkAuth = require('../middleware/check-auth');
 router.post('', checkAuth, (req, res, next) => {
     const post = new Post({
         title: req.body.title,
-        content: req.body.content
+        content: req.body.content,
+        creator: req.userData.userId
     });
+    // console.log(req.userData);
+    // return res.status(200).json({});
     post.save().then(createPost => {
         res.status(201).json({
             message: 'Post added successfully',
-            postId: createPost._id
+            post: {
+                title: createPost.title,
+                content: createPost.content,
+                id: createPost.id,
+                creator: createPost.creator
+            },
         })
     });
 });
@@ -23,10 +31,12 @@ router.put('/:id', checkAuth, (req, res, next) => {
         content: req.body.content
     });
     console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-        res.status(200).json({
-            message: 'updated successfully!'
-        });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+        if (result.nModified > 0) {
+            res.status(200).json({ message: 'updated successfully!' });
+        } else {
+            res.status(401).json({ message: 'Not authorized to update this post!' });
+        }
     });
 });
 router.get('', (req, res, next) => {
@@ -37,18 +47,20 @@ router.get('', (req, res, next) => {
         });
 });
 router.get('/:id', (req, res, next) => {
-    Post.findOne(({_id: req.params.id})).then(result => {
+    Post.findOne(({ _id: req.params.id })).then(result => {
         console.log(res);
         res.status(200).json(result);
     });
 });
 router.delete("/:id", checkAuth, (req, res, next) => {
     console.log(req.params.id);
-    Post.deleteOne({ _id: req.params.id }).then(result => {
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
         console.log(result);
-        res.status(200).json({
-            message: "Post deleted"
-        });
+        if (result.n > 0) {
+            res.status(200).json({ message: 'Post deleted successfully!' });
+        } else {
+            res.status(401).json({ message: 'Not authorized to delete this post!' });
+        }
     });
 });
